Separate the code into a library and an application.
The library is the core code, that stores and retrieves events.

The application is a demo app, with a working example of how to use
the library.

This commit only creates the structure, with a parent pom.xml and two
modules.  Next step is to move the application code from the library
module into the demo app module.